test_8

+++ 2D/training/train.py
@@ -39,7 +39,7 @@ ratio_max = ratio_potrhs(alpha, Lx, Ly)
 
 
 #Create Data
-dataset = np.load(data_dir) 
+dataset = np.load(data_dir).astype(np.float32) 
 dataloader = DataLoader(dataset, batch_size=batch_size, shuffle=True)
